requirements.txt
altgraph==0.17.4
packaging==25.0
pefile==2023.2.7
pygame-ce==2.5.5
pygbag==0.9.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.8
pywin32-ctypes==0.2.3
setuptools==80.9.0
